Move role resources to dedicated pkg
